requirements.txt
chromadb
faiss-cpu
openai
langchain
python-dotenv